requirements.txt
typer[all]==0.9.0
PyGithub==1.58.2
GitPython==3.1.32
